+++ src/commands/guild-management/addMember.ts
@@ -16,19 +16,30 @@ export const data = new SlashCommandBuilder()
             .setName("user-id")
             .setDescription("The user ID of the member to add")
             .setRequired(true)
+            .setMinLength(36)
+            .setMaxLength(36)
     )
     .addStringOption((option) =>
         option
             .setName("username")
             .setDescription("The username of the member to add")
             .setRequired(true)
-    );
+    )
+    .addStringOption((option) => {
+        option
+            .setName("player-api-token")
+            .setDescription("The API token of the player to add (optional)")
+            .setRequired(false)
+            .setMinLength(36)
+            .setMaxLength(36);
+        return option;
+    });
 
 export async function execute(interaction: ChatInputCommandInteraction) {
     await interaction.deferReply({ flags: MessageFlags.Ephemeral });
 
     const userId = interaction.options.getString("user-id");
-    if (!userId) {
+    if (!userId || !isValidUUIDv4(userId)) {
         await interaction.editReply({
             content: "Please provide a valid user ID.",
             options: {
@@ -73,7 +84,25 @@ export async function execute(interaction: ChatInputCommandInteraction) {
         return;
     }
 
-    const result = await service.updateGuildMember(userId, username, guildId);
+    const apiToken = interaction.options.getString("player-api-token");
+
+    if (apiToken && !isValidUUIDv4(apiToken)) {
+        await interaction.editReply({
+            content:
+                "Invalid API token format. Please provide a valid UUID (v4) for the player API token.",
+            options: {
+                flags: MessageFlags.Ephemeral,
+            },
+        });
+        return;
+    }
+
+    const result = await service.updateGuildMember(
+        userId,
+        username,
+        guildId,
+        apiToken ?? undefined
+    );
     if (result) {
         await interaction.editReply({
             content: `Successfully added member with ID \`${userId}\` and username \`${username}\` to the guild.`,