+++ src/models/types/TokenStatus.ts
@@ -1,0 +1,4 @@
+export interface TokenStatus {
+    count: number;
+    refreshTime: number;
+}